fix warnings

+++ vstgui/uidescription/editing/uicolorscontroller.cpp
@@ -231,8 +231,9 @@ CMouseEventResult UIColorsDataSource::dbOnMouseMoved (const CPoint& where,
 			if (description->getColor (names.at (static_cast<uint32_t> (row)).data (), color))
 			{
 				auto colorStr = color.toString ();
-				auto dropSource = CDropSource::create (colorStr.data (), colorStr.length () + 1,
-				                                       CDropSource::kText);
+				auto dropSource = CDropSource::create (
+				    colorStr.data (), static_cast<uint32_t> (colorStr.length () + 1),
+				    CDropSource::kText);
 				SharedPointer<CBitmap> dragBitmap;
 				if (auto offscreen = COffscreenContext::create (browser->getFrame (), r.getWidth (), r.getHeight ()))
 				{
